+++ src/test/java/org/folio/consortia/service/KeycloakCredentialsServiceTest.java
@@ -11,14 +11,14 @@ import org.folio.consortia.domain.dto.KeycloakTokenResponse;
 import org.folio.consortia.service.impl.KeycloakCredentialsServiceImpl;
 import org.folio.consortia.support.CopilotGenerated;
 import org.folio.tools.store.SecureStore;
-import org.folio.tools.store.exception.NotFoundException;
+import org.folio.tools.store.exception.SecretNotFoundException;
+import org.folio.tools.store.properties.SecureStoreProperties;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.beans.factory.annotation.Value;
 
 import java.util.List;
 
@@ -36,13 +36,12 @@ class KeycloakCredentialsServiceTest {
   private KeycloakLoginClientProperties keycloakClientProperties;
   @Mock
   private SecureStore secureStore;
+  @Mock
+  private SecureStoreProperties secureStoreProperties;
 
   @InjectMocks
   @Spy
   private KeycloakCredentialsServiceImpl keycloakCredentialsService;
-
-  @Value("${folio.environment}")
-  private String folioEnvironment;
 
   @Test
   void getClientCredentials_returnsValidCredentials() {
@@ -84,26 +83,30 @@ class KeycloakCredentialsServiceTest {
   void getMasterAuthToken_returnsValidToken() {
     String clientId = "clientId";
     String clientSecret = "clientSecret";
+    var secureStoreEnv = "folio";
     var tokenResponse = createTokenResponse();
 
     when(keycloakProperties.getClientId()).thenReturn(clientId);
     when(secureStore.get(anyString())).thenReturn(clientSecret);
     when(keycloakClient.login(anyMap())).thenReturn(tokenResponse);
+    when(secureStoreProperties.getEnvironment()).thenReturn(secureStoreEnv);
 
     String token = keycloakCredentialsService.getMasterAuthToken();
 
     assertEquals(AUTH_TOKEN, token);
-    verify(secureStore).get("%s_%s_%s".formatted(folioEnvironment, "master", clientId));
+    verify(secureStore).get("%s_%s_%s".formatted(secureStoreEnv, "master", clientId));
   }
 
   @Test
   void getMasterAuthToken_throwsException_secretNotFound() {
     String clientId = "clientId";
+    var secureStoreEnv = "folio";
     when(keycloakProperties.getClientId()).thenReturn(clientId);
-    when(secureStore.get(anyString())).thenThrow(new NotFoundException("Not found"));
+    when(secureStore.get(anyString())).thenThrow(new SecretNotFoundException("Not found"));
+    when(secureStoreProperties.getEnvironment()).thenReturn(secureStoreEnv);
 
     assertThrows(IllegalStateException.class, () -> keycloakCredentialsService.getMasterAuthToken());
-    verify(secureStore).get("%s_%s_%s".formatted(folioEnvironment, "master", clientId));
+    verify(secureStore).get("%s_%s_%s".formatted(secureStoreEnv, "master", clientId));
   }
 
   private static KeycloakTokenResponse createTokenResponse() {